+++ src/brasil/gov/portal/tests/test_viewlets.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from brasil.gov.portal.browser.viewlets.content import DocumentBylineViewlet
-from brasil.gov.portal.browser.viewlets.destaques import Destaques_Viewlet
 from brasil.gov.portal.browser.viewlets.logo import LogoViewlet
 from brasil.gov.portal.browser.viewlets.nitf_byline import NITFBylineViewlet
 from brasil.gov.portal.browser.viewlets.redes import RedesSociaisViewlet
@@ -8,7 +7,6 @@ from brasil.gov.portal.browser.viewlets.servicos import ServicosViewlet
 from brasil.gov.portal.testing import INTEGRATION_TESTING
 from plone import api
 from plone.app.layout.globals.interfaces import IViewView
-from plone.app.testing import logout
 from zope.interface import implementer
 
 import unittest
@@ -26,81 +24,6 @@ def esconde_data():
     record = \
         'brasil.gov.portal.controlpanel.portal.ISettingsPortal.esconde_data'
     api.portal.set_registry_record(record, True)
-
-
-class DestaquesViewletTestCase(unittest.TestCase):
-
-    layer = INTEGRATION_TESTING
-
-    def setUp(self):
-        self.portal = self.layer['portal']
-        self.request = self.layer['request']
-        with api.env.adopt_roles(['Manager']):
-            self.destaques = api.content.create(
-                type='collective.cover.content',
-                container=self.portal,
-                id='destaques',
-                title=u'Destaques',
-            )
-
-    def viewlet(self, context=None):
-        context = context or self.portal
-        viewlet = Destaques_Viewlet(context, self.request, None, None)
-        viewlet.update()
-        return viewlet
-
-    def test_available(self):
-        viewlet = self.viewlet()
-        self.assertTrue(viewlet.available())
-
-    def test_not_available_on_folder(self):
-        with api.env.adopt_roles(['Manager']):
-            api.content.create(
-                type='Folder',
-                container=self.portal,
-                id='pasta',
-                title=u'Uma pasta',
-            )
-        viewlet = self.viewlet(self.portal['pasta'])
-        self.assertFalse(viewlet.available())
-
-    def test_not_available(self):
-        with api.env.adopt_roles(['Manager']):
-            # Apagamos a capa de destaques
-            api.content.delete(obj=self.portal['destaques'])
-        viewlet = self.viewlet()
-        self.assertFalse(viewlet.available())
-
-    def test_available_for_different_content_type(self):
-        with api.env.adopt_roles(['Manager']):
-            # Apagamos a capa de destaques
-            api.content.delete(obj=self.portal['destaques'])
-            # Colocamos uma pasta no mesmo lugar
-            pasta = api.content.create(
-                type='Folder',
-                container=self.portal,
-                id='destaques',
-                title=u'Uma pasta com destaques',
-            )
-        viewlet = self.viewlet()
-        # O Viewlet deve detectar o problema e nao exibir nada
-        self.assertEqual(pasta.portal_type, 'Folder')
-        self.assertFalse(viewlet.available())
-
-    def test_editable(self):
-        viewlet = self.viewlet()
-        self.assertTrue(viewlet.editable())
-
-    def test_not_editable_by_anonymous(self):
-        logout()
-        viewlet = self.viewlet()
-        self.assertFalse(viewlet.editable())
-
-    def test_portal_url(self):
-        logout()
-        viewlet = self.viewlet()
-        self.assertEqual(viewlet.portal_url,
-                         self.portal.portal_url())
 
 
 class LogoViewletTestCase(unittest.TestCase):